mod: 移动universal依赖的逻辑到单独的文件里
